requirements.txt
requests
beautifulsoup4
pyyaml
matplotlib
Flask
aiohttp  # worker 异步 HTTP 客户端